Create a GitHub page to display results
Fixes #3. There is also a gh-page branch to host the site. The site will
need to also live in the master branch as it will be created in part
through the application which will write stuff needed for the site's
creation. The site will then need to be copied over to the correct
branch. A shell script (gensite.sh) will take automate all of that
including the site's creation and deployment.